Simplify migration preflight scanner
Collapse the unused preflight strategy abstraction back into a single scanner now that ripgrep is no longer used. Keep the ambiguous-marker scan behavior covered by the existing migration tests.
