+++ Assets/Tests/Editor/ThirdPartyToolMigrationFileServiceTests.cs
@@ -126,11 +126,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         }
 
         [Test]
-        public void ManagedPreflightSearchStrategy_WhenSourceHasNoMigrationMarkers_ReturnsNoTargets()
+        public void PreflightScanner_WhenSourceHasNoMigrationMarkers_ReturnsNoTargets()
         {
             // Verifies that startup preflight can skip full scans when no migration marker text exists.
             MigrationTargetPreflightResult result =
-                ThirdPartyToolMigrationManagedPreflightSearchStrategy.InspectSourceText(
+                ThirdPartyToolMigrationPreflightScanner.InspectSourceText(
                     "public sealed class PlainTool {}",
                     ".cs");
 
@@ -138,11 +138,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         }
 
         [Test]
-        public void ManagedPreflightSearchStrategy_WhenLegacyToolSourceExists_ReturnsHasTargets()
+        public void PreflightScanner_WhenLegacyToolSourceExists_ReturnsHasTargets()
         {
             // Verifies that startup preflight detects direct legacy custom tool source immediately.
             MigrationTargetPreflightResult result =
-                ThirdPartyToolMigrationManagedPreflightSearchStrategy.InspectSourceText(
+                ThirdPartyToolMigrationPreflightScanner.InspectSourceText(
                     "using io.github.hatayama.uLoopMCP; [McpTool] public sealed class HelloTool {}",
                     ".cs");
 
@@ -150,11 +150,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         }
 
         [Test]
-        public void ManagedPreflightSearchStrategy_WhenLegacyNamespaceIsOnlyComment_ReturnsNeedsFullScan()
+        public void PreflightScanner_WhenLegacyNamespaceIsOnlyComment_ReturnsNeedsFullScan()
         {
             // Verifies that startup preflight does not report targets from marker text inside comments.
             MigrationTargetPreflightResult result =
-                ThirdPartyToolMigrationManagedPreflightSearchStrategy.InspectSourceText(
+                ThirdPartyToolMigrationPreflightScanner.InspectSourceText(
                     "// using io.github.hatayama.uLoopMCP;",
                     ".cs");
 
@@ -162,11 +162,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         }
 
         [Test]
-        public void ManagedPreflightSearchStrategy_WhenLegacyAsmdefReferenceExists_ReturnsHasTargets()
+        public void PreflightScanner_WhenLegacyAsmdefReferenceExists_ReturnsHasTargets()
         {
             // Verifies that startup preflight detects legacy asmdef references without building an inventory.
             MigrationTargetPreflightResult result =
-                ThirdPartyToolMigrationManagedPreflightSearchStrategy.InspectSourceText(
+                ThirdPartyToolMigrationPreflightScanner.InspectSourceText(
                     @"{ ""references"": [ ""uLoopMCP.Editor"" ] }",
                     ".asmdef");
 
@@ -174,7 +174,7 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
         }
 
         [Test]
-        public async Task ManagedPreflightSearchStrategy_WhenAmbiguousFileExistsWithDirectTarget_ReturnsHasTargets()
+        public async Task PreflightScanner_WhenAmbiguousFileExistsWithDirectTarget_ReturnsHasTargets()
         {
             // Verifies that an ambiguous marker does not force full scan before later direct targets are checked.
             string projectRoot = CreateProjectRoot();
@@ -188,11 +188,11 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
                 File.WriteAllText(
                     Path.Combine(toolDirectory, "1_HelloTool.cs"),
                     "using io.github.hatayama.uLoopMCP; [McpTool] public sealed class HelloTool {}");
-                ThirdPartyToolMigrationManagedPreflightSearchStrategy strategy =
-                    new ThirdPartyToolMigrationManagedPreflightSearchStrategy();
 
                 MigrationTargetPreflightResult result =
-                    await strategy.FindMigrationTargetAsync(projectRoot, CancellationToken.None);
+                    await ThirdPartyToolMigrationPreflightScanner.FindMigrationTargetAsync(
+                        projectRoot,
+                        CancellationToken.None);
 
                 Assert.That(result, Is.EqualTo(MigrationTargetPreflightResult.HasTargets));
             }
@@ -200,54 +200,6 @@ namespace io.github.hatayama.UnityCliLoop.Tests.Editor
             {
                 Directory.Delete(projectRoot, recursive: true);
             }
-        }
-
-        [Test]
-        public async Task PreflightSearchPipeline_WhenFirstStrategyIsDecisive_DoesNotRunNextStrategy()
-        {
-            // Verifies that rg and managed preflight strategies share one decisive-result pipeline.
-            FakePreflightSearchStrategy firstStrategy =
-                new FakePreflightSearchStrategy(MigrationTargetPreflightResult.NoTargets);
-            FakePreflightSearchStrategy secondStrategy =
-                new FakePreflightSearchStrategy(MigrationTargetPreflightResult.HasTargets);
-            ThirdPartyToolMigrationPreflightSearchPipeline pipeline =
-                new ThirdPartyToolMigrationPreflightSearchPipeline(
-                    new IThirdPartyToolMigrationPreflightSearchStrategy[]
-                    {
-                        firstStrategy,
-                        secondStrategy
-                    });
-
-            MigrationTargetPreflightResult result =
-                await pipeline.FindMigrationTargetAsync("ProjectRoot", CancellationToken.None);
-
-            Assert.That(result, Is.EqualTo(MigrationTargetPreflightResult.NoTargets));
-            Assert.That(firstStrategy.CallCount, Is.EqualTo(1));
-            Assert.That(secondStrategy.CallCount, Is.EqualTo(0));
-        }
-
-        [Test]
-        public async Task PreflightSearchPipeline_WhenFirstStrategyNeedsFullScan_RunsNextStrategy()
-        {
-            // Verifies that inconclusive preflight results fall through to the next strategy.
-            FakePreflightSearchStrategy firstStrategy =
-                new FakePreflightSearchStrategy(MigrationTargetPreflightResult.NeedsFullScan);
-            FakePreflightSearchStrategy secondStrategy =
-                new FakePreflightSearchStrategy(MigrationTargetPreflightResult.HasTargets);
-            ThirdPartyToolMigrationPreflightSearchPipeline pipeline =
-                new ThirdPartyToolMigrationPreflightSearchPipeline(
-                    new IThirdPartyToolMigrationPreflightSearchStrategy[]
-                    {
-                        firstStrategy,
-                        secondStrategy
-                    });
-
-            MigrationTargetPreflightResult result =
-                await pipeline.FindMigrationTargetAsync("ProjectRoot", CancellationToken.None);
-
-            Assert.That(result, Is.EqualTo(MigrationTargetPreflightResult.HasTargets));
-            Assert.That(firstStrategy.CallCount, Is.EqualTo(1));
-            Assert.That(secondStrategy.CallCount, Is.EqualTo(1));
         }
 
         [Test]
@@ -4914,25 +4866,5 @@ public sealed class ScreenshotResponse : UnityCliLoopToolResponse
             }
         }
 
-        private sealed class FakePreflightSearchStrategy : IThirdPartyToolMigrationPreflightSearchStrategy
-        {
-            private readonly MigrationTargetPreflightResult _result;
-
-            public FakePreflightSearchStrategy(MigrationTargetPreflightResult result)
-            {
-                _result = result;
-            }
-
-            public int CallCount { get; private set; }
-
-            public Task<MigrationTargetPreflightResult> FindMigrationTargetAsync(
-                string projectRoot,
-                CancellationToken ct)
-            {
-                // Verifies pipeline sequencing without depending on the filesystem or rg.
-                CallCount++;
-                return Task.FromResult(_result);
-            }
-        }
     }
 }
