--- Packages/src/Editor/Infrastructure/ThirdPartyToolMigration/ThirdPartyToolMigrationManagedPreflightSearchStrategy.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 7280a38a11be4f0eb1105f7b99546c28
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData:
-  assetBundleName:
-  assetBundleVariant:
